requirements.txt
Pillow
colorsys